requirements.txt
aiofiles
aiohttp
anyio
attrs
bcrypt
certifi
cffi
charset-normalizer
click
cryptography
fastapi
Flask
google-api-core
google-api-python-client
google-auth
grpcio
httpx
huggingface-hub
Jinja2
joblib
kubernetes
langchain
lxml
matplotlib
numpy
openai
pandas
pillow
protobuf
pydantic
pytesseract
python-dotenv
PyYAML
requests
scikit-learn
scipy
SQLAlchemy
starlette
sympy
torch
transformers
tqdm
uvicorn
Werkzeug
PyPDF2
docx
